+++ projects/core/src/main/kotlin/properties-jvm.kt
@@ -25,17 +25,17 @@ import java.util.concurrent.atomic.AtomicInteger
 import kotlin.properties.ReadWriteProperty
 
 
-@suppress("UNCHECKED_CAST")
+@Suppress("UNCHECKED_CAST")
 public class ThreadSafeLazyVar<T>(initializer: () -> T) : ReadWriteProperty<Any?, T> {
-    private volatile var threadCount: AtomicInteger? = AtomicInteger(0)
-    private volatile var initializer: (() -> T)?
-    private volatile var value: Any? = null
+    private @Volatile var threadCount: AtomicInteger? = AtomicInteger(0)
+    private @Volatile var initializer: (() -> T)?
+    private @Volatile var value: Any? = null
 
     init {
         this.initializer = initializer
     }
 
-    public override fun get(thisRef: Any?, desc: PropertyMetadata): T {
+    public override fun get(thisRef: Any?, property: PropertyMetadata): T {
         //Busy /Spin lock, expect quick initialization
         while (value == null) {
             val counter = threadCount
@@ -56,23 +56,23 @@ public class ThreadSafeLazyVar<T>(initializer: () -> T) : ReadWriteProperty<Any?
         return unmask(value) as T
     }
 
-    public override fun set(thisRef: Any?, desc: PropertyMetadata, value: T) {
+    public override fun set(thisRef: Any?, property: PropertyMetadata, value: T) {
         this.value = mask(value)
     }
 
     val initialized: Boolean get() = value != null
 }
 
-@suppress("UNCHECKED_CAST")
+@Suppress("UNCHECKED_CAST")
 public class TrackChangesVar<T>(private val source: () -> T) : ReadWriteProperty<Any?, T> {
-    private volatile var value: Any? = null
+    private @Volatile var value: Any? = null
 
-    public override fun get(thisRef: Any?, desc: PropertyMetadata): T {
+    public override fun get(thisRef: Any?, property: PropertyMetadata): T {
         val curVal = value
         return if (curVal != null) unmask(curVal) as T else source()
     }
 
-    public override fun set(thisRef: Any?, desc: PropertyMetadata, value: T) {
+    public override fun set(thisRef: Any?, property: PropertyMetadata, value: T) {
         this.value = mask(value)
     }